workers/pyproject.toml
[tool.poetry]
name = "workers"
version = "0.1.0"
description = ""
authors = ["Захаров Павел <pzakharov83@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.1"
weasyprint = "^57.2"
jinja2 = "^3.1.2"
python-barcode = {extras = ["images"], version = "^0.14.0"}
qrcode = {extras = ["pil"], version = "^7.3.1"}
sqlalchemy = "^1.4.46"
pyodbc = "^4.0.35"


[tool.poetry.group.dev.dependencies]
sqlalchemy = {extras = ["mypy"], version = "^1.4.46"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
